Create README.md
The Gamja means the potato. The potatoflower font is characterized by the handwriting style of the potato flower with the feminine shape of the potato flower, and it gives a vivid impression and its characteristics of the soft cord and tie beams
As like yellow potato flowers add soft beauty and gorgeous feelings. Thanks to these potato flowers are likely to bring good luck to me.